fix: move NZ truth seeding to application-scoped singleton
viewModelScope was cancelled on navigation, leaving the SharedPrefs
flag unset and causing an infinite delete-and-reseed loop.

NzTruthSeedingService owns its own CoroutineScope (SupervisorJob +
Dispatchers.IO) so seeding survives ViewModel recreation and completes
exactly once per seed-guard version. ChatViewModel.seedKiwiTruthsIfNeeded()
removed; replaced with nzTruthSeedingService.seedIfNeeded().
